AI_chat/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
google-generativeai==0.8.5
python-multipart==0.0.20
python-dotenv==1.0.0
requests==2.32.3
cryptography==42.0.5
beautifulsoup4==4.12.3
